requirements.txt
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
cfgv==3.4.0
click==8.0.1
colorama==0.4.6
distlib==0.4.0
Django==4.2.23

djangorestframework==3.16.1
drf-spectacular==0.28.0
filelock==3.19.1
gunicorn==23.0.0
identify==2.6.13
inflection==0.5.1
isort==6.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pre_commit==4.3.0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
virtualenv==20.34.0
